reorder decorators to add back namespace
